requirements.txt
greenlet==3.0.3
pyaes==1.6.1
Pyrogram==2.0.106
PySocks==1.7.1
python-dotenv==1.0.1
SQLAlchemy==2.0.27
TgCrypto==1.2.5
typing-extensions==4.9.0
